extract stories into folder
