+++ src/core/components/radio/stories/horizontal.tsx
@@ -1,0 +1,14 @@
+import React from "react"
+import { RadioGroup, Radio } from "../index"
+
+const horizontal = () => (
+	<RadioGroup orientation="horizontal" name="yes-or-no">
+		<Radio value="yes" label="Yes" />
+		<Radio value="no" label="No" checked={true} />
+	</RadioGroup>
+)
+horizontal.story = {
+	name: "orientation horizontal",
+}
+
+export { horizontal }
